Add initial test harness (#1)
This commit adds the baseline for all kinds of unit testing for the ygopen ecosystem; At the moment the main building blocks are tested: Value wrappers and Frame objects, which are the containers used throughout the message parsing process and reused by other projects.

There's still a lot more to unit test:
  * The Board object which contains the frame object plus other data pertaining to an actual duel.
  * The protobuf parsers, both the message parser and the query parser, as well as the undoing parsers which are not yet completed.
  * The encoder and decoder, which requires much more involvement as it has several dozens different messages to parse and decode.
  * Death tests for expected failures when using objects and interfaces in a improper manner, this just requires more work and it's probably coming soon.

Nevertheless, this is a massive step in ensuring the stability and re-usability of the code written, which is very nice. Turns out that adding a test suite wasn't so hard after all!