use `__has_include` and cast results of popcnt to `unsigned`
the feature macro `__cpp_lib_bitops` *is defined* within <bits>, so it is
useless to attempt checking for it before that.

I got no idea why all popcnt implementations return int rather than an unsigned
to begin with, but it feels safer to have an unsigned regardless (kinda like
size_t for sizes).